pyproject.toml
[tool.poetry]
name = "space_telegram"
version = "0.1.0"
description = ""
authors = ["Danil Nikitin <d.nikitin.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
python-telegram-bot = "13.0"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
pathvalidate = "^2.5.0"
click = "^8.1.3"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.971"
types-requests = "^2.28.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
